refactor: factor out backdrop
